+++ src/builder.ts
@@ -141,8 +141,18 @@ import {
   NexusObjectTypeExtension,
   NexusInputObjectTypeExtension,
 } from "./extensions";
+import * as Plugins from "./plugins-current";
 
 export type Maybe<T> = T | null;
+
+export type NexusAcceptedTypeDef =
+  | AllNexusNamedTypeDefs
+  | NexusExtendInputTypeDef<string>
+  | NexusExtendTypeDef<string>
+  | GraphQLNamedType
+  | DynamicInputMethodDef<string>
+  | DynamicOutputMethodDef<string>
+  | DynamicOutputPropertyDef<string>;
 
 type NexusShapedOutput = {
   name: string;
@@ -259,7 +269,7 @@ export interface BuilderConfig {
    * List of plugins to apply to Nexus, with before/after hooks
    * executed first to last: before -> resolve -> after
    */
-  plugins?: PluginDef[];
+  plugins?: Plugins.PluginDef[];
 }
 
 export type SchemaConfig = BuilderConfig & {
@@ -334,6 +344,12 @@ export type BuilderLens = {
  * circular references at this step, while fields will guard for it during lazy evaluation.
  */
 export class SchemaBuilder {
+  /**
+   * Used to track all _GraphQL_ types that have been added to the builder.
+   * This supports hasType method which permits asking the question "Will
+   * the GraphQL schema have _this_ type (name)".
+   */
+  protected allTypeDefs: Record<string, NexusAcceptedTypeDef> = {};
   /**
    * Used to check for circular references.
    */
@@ -464,10 +480,6 @@ export class SchemaBuilder {
     return this.config;
   }
 
-  getPlugins(): PluginDef[] {
-    return this.plugins;
-  }
-
   /**
    * Add type takes a Nexus type, or a GraphQL type and pulls
    * it into an internal "type registry". It also does an initial pass
@@ -514,6 +526,10 @@ export class SchemaBuilder {
         return;
       }
       throw extendError(typeDef.name);
+    }
+
+    if (!this.allTypeDefs[typeDef.name]) {
+      this.allTypeDefs[typeDef.name] = typeDef;
     }
 
     if (isNexusScalarTypeDef(typeDef) && typeDef.value.asNexusMethod) {
@@ -1430,7 +1446,14 @@ export function buildTypesInternal<
   TypeMapDefs extends Record<string, GraphQLNamedType> = any
 >(types: any, config: BuilderConfig): BuildTypes<TypeMapDefs> {
   const builder = new SchemaBuilder(config);
+  const plugins = config.plugins || [];
+  const pluginControllers = plugins.map((plugin) =>
+    Plugins.initialize(builder, plugin)
+  );
   addTypes(builder, types);
+  pluginControllers.forEach((pluginController) =>
+    pluginController.triggerOnInstall()
+  );
   return builder.getFinalTypeMap();
 }